set map_scu/ipm line buffer

+++ src/com/ipred.rs
@@ -16,13 +16,13 @@ pub(crate) fn evc_get_mpm_b(
         && map_scu[(scup - 1) as usize].GET_IF() != 0
         && map_scu[(scup - 1) as usize].GET_COD() != 0
     {
-        ipm_l = map_ipm[(scup - 1) as usize] as usize + 1;
+        ipm_l = (map_ipm[(scup - 1) as usize] as i8 + 1) as usize;
     }
     if y_scu > 0
         && map_scu[(scup - w_scu as u32) as usize].GET_IF() != 0
         && map_scu[(scup - w_scu as u32) as usize].GET_COD() != 0
     {
-        ipm_u = map_ipm[(scup - w_scu as u32) as usize] as usize + 1;
+        ipm_u = (map_ipm[(scup - w_scu as u32) as usize] as i8 + 1) as usize;
     }
 
     &evey_tbl_mpm[ipm_l as usize][ipm_u as usize]
